requirements.txt
numpy
matplotlib
seaborn
PyYAML==6.0.1
matplotlib==3.8.1
numpy==1.26.1